requirements.txt
amqp==5.0.1
asgiref==3.3.0
billiard==3.6.3.0
celery==5.0.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
cryptography==3.1.1
defusedxml==0.7.0rc1
Django==3.1.2
django-audiofield==0.10.0
django-celery==3.3.1
django-ckeditor==6.0.0
django-crispy-forms==1.9.2
django-embed-video==1.3.3
django-etc==1.2.0
django-hitcount==1.3.3
django-js-asset==1.2.2
idna==2.10
kombu==5.0.2
oauthlib==3.1.0
Pillow==7.2.0
prompt-toolkit==3.0.8
pycparser==2.20
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.25.10
vine==5.0.0
wcwidth==0.2.5
